+++ src/constants.js
@@ -23,12 +23,7 @@ const TYPES = {
     qs: 'qs',
 };
 
-const USER_AGENT = platform() === 'win32'
-    ? 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.104 Safari/537.36'
-    : 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.150 Safari/537.36';
-
 module.exports = {
     LABELS,
-    USER_AGENT,
     TYPES,
 };